requirements.txt
kafka-python
pyspark==3.5.0
delta-spark==3.0.0
pandas
google-api-python-client
python-dotenv
requests
textblob
langdetect
streamlit
plotly
statsmodels
dotenv 
requests 
pillow
opencv-python
deltalake